Selenium方法改为Requests方法: 获取EPC列表部分完成
