+++ .gitignore
@@ -16,4 +16,5 @@ dist/
 *.bat
 
 # User config file
+/.vscode
 config.json
