+++ main.py
@@ -1,0 +1,7 @@
+from tkinter import *
+from gui import *
+
+
+root = Tk()
+GUI(root)
+root.mainloop()
