--- main_gui.py
@@ -1,6 +1,0 @@
-from tkinter import *
-from graphic import *
-
-root = Tk()
-GUI(root)
-root.mainloop()
